Adding angle_and_depth funcction and tests.
Then angle_and_depth function returns the angle of the n-phenomenon in
degrees as it destroys the entropy generated by rule 30, and the maximum
depth of the central column of the automata before it is destroy by the
n-phenomenon.

The function can take a very long time to run, so it's results are
cached in a csv file which is accessed via the panda library.